add CMS live controller and move GA code
